refactor(xlsx): extract drawing and slicer modules from parser lib.rs
- drawing.rs (~790 lines): twoCellAnchor parsing, xfrm, custom paths,
  txBody/shape geometry, embedded image data URLs, shape groups.
- slicer.rs (~290 lines): pivot cache fields, slicer caches, slicer
  definitions and anchors (Office 2010+ extension).

Shared zip helpers (read_zip_bytes, mime_from_ext) are now pub(crate).
lib.rs drops from 2674 to 1611 lines (72% smaller than the original
5804). Verified behavior-invariant via VRT regression (65/65, 0.0% diff).
